MOSU-40 feat: 회원가입 로그인 기능 구현 (#50)
* MOSU-5 feat: 로그인 Service 구현

* MOSU-5 feat: 로그인 Controller 구현

* MOSU-5 fix: applicartion 패키지로 이동

* MOSU-5 feat: 로그인 입력값 검증

* MOSU-5 refactor: 패키지명 수정

* MOSU-40 feat 회원가입 Controller 구현

* MOSU-40 feat 회원가입 Service 구현

* MOSU-40 feat PENDING ROLE 추가

* MOSU-40 feat EncodeUtil 구현

* MOSU-40 feat 다날 인증 인터페이스 구현

* MOSU-40 feat: WebMvcConfig 설정

* MOSU-40 feat: UserId 추출 어노테이션 구현

* MOSU-40 feat: RefreshToken 구현

* MOSU-40 feat: Profile User id 적용

* MOSU-40 feat: ErrorCode 처리 추가

* MOSU-40 feat: reissue url 열기

* MOSU-40 feat: security 설정

* MOSU-40 refactor: 도메인 패키지 이동

* MOSU-40 refactor: 회원가입 생성 책임을 dto로 위임

* MOSU-40 refactor: 추론 타입을 void로 변경

* MOSU-40 refactor: reissue 상수 분리

* MOSU-40 refactor: PrincipalDetails null 제거

* MOSU-40 fix: @Transactional 제거

* MOSU-40 test: e2e 테스트

+++ src/main/java/life/mosu/mosuserver/application/auth/AuthTokenManager.java
@@ -2,6 +2,8 @@ package life.mosu.mosuserver.application.auth;
 
 import jakarta.servlet.http.HttpServletRequest;
 import life.mosu.mosuserver.domain.user.UserJpaEntity;
+import life.mosu.mosuserver.global.exception.CustomRuntimeException;
+import life.mosu.mosuserver.global.exception.ErrorCode;
 import life.mosu.mosuserver.presentation.auth.dto.Token;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
@@ -25,13 +27,21 @@ public class AuthTokenManager {
 
     public Token reissueAccessToken(final HttpServletRequest servletRequest) {
         final String refreshTokenString = refreshTokenService.resolveToken(servletRequest);
+        if (refreshTokenString == null) {
+            throw new CustomRuntimeException(ErrorCode.NOT_FOUND_REFRESH_TOKEN);
+        }
 
-        final Authentication authentication = refreshTokenService.getAuthentication(refreshTokenString);
+        final Authentication authentication = refreshTokenService.getAuthentication(
+                refreshTokenString);
         final PrincipalDetails principalDetails = (PrincipalDetails) authentication.getPrincipal();
         final UserJpaEntity user = principalDetails.user();
 
         refreshTokenService.deleteRefreshToken(user.getId());
 
         return generateAuthToken(user);
+    }
+
+    public Long getAccessTokenExpireTime() {
+        return accessTokenService.expireTime;
     }
 }
